Functions applied to dictionary arrays decode the output array.
When pyarrow doesn't support a function on a dictionary array, the function is applied to the dictionary itself, as an optimization.
However this may result in duplicate dictionary entries, which breaks some dictionary operations, e.g., `unique`.
So the output must be decoded anyway, but is still faster than applying the function to all scalars.