Flatten grouping on multiple arrays by creating a unique composite key.
This is significantly faster for nested grouping in the case of many penultimate groups.
It can be slower for extremely large number of final groups, due to cache locality on hash tables.
The encoding itself is a little hacky and in theory could overflow, but will likely be replaced when ARROW-3978 is implemented.